requirements.txt
torch
matplotlib
torchvision
hydra-core==1.3.2